requirements.txt
fastapi
text2emotion
dotenv
requests
uvicorn
mysql.connector